+++ packages/mobile/src/components/ImageView/Spotlight.tsx
@@ -2,15 +2,18 @@ import React, { useContext, useState } from 'react'
 import { deepEqual, onUpdate, ReactState, TypeGuards, usePrevious, useUnmount } from '@codeleap/common'
 
 import uuid from 'react-native-uuid'
-import { ImageView } from './component'
+import { ImageView, ImageViewProps } from './component'
 import { ImageProps } from '../Image'
 import { ImageURISource, ImageRequireSource } from 'react-native'
+import { View } from '../View'
+import { Text } from '../Text'
+
 type ImageSource = ImageURISource | ImageRequireSource
 
 type TImage = {
-    source: ImageSource
-    created: number
-    id: string
+  source: ImageSource
+  created: number
+  id: string
 
 }
 type ImageList = Record<string, TImage>
@@ -18,19 +21,18 @@ type ImageList = Record<string, TImage>
 type SpotlightState = ReactState<Record<string, ImageList>>
 type IndexState = ReactState<Record<string, number>>
 type TSpotlightCtx = {
-    spotlights: SpotlightState[0]
-    setSpotlights: SpotlightState[1]
-    indexes: IndexState[0]
-    setIndexes: IndexState[1]
-
+  spotlights: SpotlightState[0]
+  setSpotlights: SpotlightState[1]
+  indexes: IndexState[0]
+  setIndexes: IndexState[1]
 }
 
 const SpotlightCtx = React.createContext({} as TSpotlightCtx)
 
-export const SpotlightProvider:React.FC<React.PropsWithChildren<any>> = ({ children }) => {
+export const SpotlightProvider: React.FC<React.PropsWithChildren<any>> = ({ children }) => {
   const [spotlights, setSpotlights] = useState<TSpotlightCtx['spotlights']>({})
   const [indexes, setIndexes] = useState<TSpotlightCtx['indexes']>({})
-  const ctxValue:TSpotlightCtx = {
+  const ctxValue: TSpotlightCtx = {
     spotlights,
     setSpotlights,
     indexes,
@@ -143,17 +145,49 @@ export const useImageSpotlight = (name: string | null, src: ImageProps['source']
   }
 }
 
-export const Spotlight = ({ name }: {name: string}) => {
+export type SpotlightHeaderComponent = {
+  imageIndex: number
+  spotlight: ReturnType<typeof useSpotlight>
+}
+
+export type SpootlightFooterComponent = React.ComponentType<{
+  imageIndex: number
+  imagesLength: number
+  spotlight: ReturnType<typeof useSpotlight>
+}>
+
+type FooterComponentProps = SpotlightHeaderComponent
+
+export type SpotlightProps = {
+  name?: string
+  HeaderComponent?: (props: SpotlightHeaderComponent) => JSX.Element
+  FooterComponent?: (props: FooterComponentProps) => JSX.Element
+  showFooter?: boolean
+} & ImageViewProps
+
+const DefaultFooterComponent: SpootlightFooterComponent = ({ imageIndex, imagesLength }) => (
+  <View variants={['marginBottom:5', 'alignCenter']}>
+    <Text text={imageIndex + 1 + '/' + imagesLength} />
+  </View>
+)
+
+export const Spotlight = (props: SpotlightProps) => {
+  const { name, HeaderComponent, showFooter, FooterComponent, ...rest } = props
   const spotlight = useSpotlight(name)
+
   useUnmount(() => {
     spotlight.clear()
   })
+
+  const Footer = showFooter ? FooterComponent || DefaultFooterComponent : (() => null)
   return <ImageView
     imageIndex={spotlight.currentIndex}
     images={spotlight.images.map(x => x.source)}
     keyExtractor={(_, index) => index.toString()}
     onRequestClose={spotlight.close}
     visible={typeof spotlight.currentIndex !== 'undefined'}
-
+    {...rest}
+    FooterComponent={({ imageIndex }) => <Footer imageIndex={imageIndex} spotlight={spotlight} imagesLength={spotlight.images.length} />}
+    HeaderComponent={!!HeaderComponent ? ({ imageIndex }) => <HeaderComponent spotlight={spotlight} imageIndex={imageIndex} /> : undefined}
   />
 }