+++ packages/mobile/src/components/ImageView/component.tsx
@@ -3,23 +3,10 @@ import * as React from 'react'
 import { onUpdate, PropsOf } from '@codeleap/common'
 import _ImageView from 'react-native-image-viewing'
 import { StatusBar } from 'react-native'
-import { View } from '../View'
-import { Text } from '../Text'
-
-type FooterComponentType = React.ComponentType<{
-  imageIndex: number
-  imagesLength: number
-}>
-
-const FooterComponent: FooterComponentType = ({ imageIndex, imagesLength }) => (
-  <View variants={['marginBottom:5', 'alignCenter']}>
-    <Text text={imageIndex + 1 + '/' + imagesLength}/>
-  </View>
-)
 
 export type ImageViewProps = PropsOf<typeof _ImageView>
 
-export const ImageView: React.FC<ImageViewProps> = (props) => {
+export const ImageView = (props: ImageViewProps) => {
   onUpdate(() => {
     StatusBar.setHidden(props.visible)
   }, [props.visible])
@@ -27,7 +14,6 @@ export const ImageView: React.FC<ImageViewProps> = (props) => {
   return (
     <_ImageView
       doubleTapToZoomEnabled={false}
-      FooterComponent={({ imageIndex }) => <FooterComponent imageIndex={imageIndex} imagesLength={props.images.length}/>}
       presentationStyle={'overFullScreen'}
       {...props}
     />