+++ packages/web/src/components/Select/types.ts
@@ -93,6 +93,10 @@ export type SelectProps<T = any, Multi extends boolean = false> = React.PropsWit
     itemProps?: ButtonProps
     loadingIndicatorSize?: number
     limit?: number
+    loadInitialValue?: boolean
+    loadingMessage?: string
+    selectedOption?: ReactSelectProps<T>['value']
+    setSelectedOption?: ReactSelectProps<T>['onValueChange']
   } & Omit<
     ReactSelectProps<T, Multi>,
     'isSearchable' | 'isClearable' | 'isDisabled' | 'loadingMessage' | 'filterOption' |